Formatting tweaks
I think someone forgot they were writing Markdown, not Wikicode, because `'''technical standpoint'''` doesn't mean anything in Markdown but it was sure confusing the parser trying to deal with it. I changed it to `**technical standpoint**` which will bold it as presumably intended.
I also changed all of the examples to quoted text, instead of fenced code blocks, and wrapped code blocks **only** around the actual code in the examples. Because, it strikes me that it _would_ be a bad example to mix code and prose indiscriminately in a code block, the way the 'Tell the user...' "good" example previously did.